requirements.txt
pydantic>=2.0.0
pytest>=7.0.0
tqdm>=4.0.0
openai>=1.0.0
anthropic>=0.18.0
PyYAML>=6.0
python-dotenv>=1.0.0
